Move authenticator instance to a single package and avoid passing it in parameters.
